Refactor workspace hierarchy validator into purpose-based files
